+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Cbanges
 
 * Add association name options
+* Modify not to define unnecessary class methods for ActiveRecord model
 
 
 ## 0.1.1 (2018/09/19)